1. Replace string related function with strtool 2. Replace mysql connector c++ with connect c 6.1

+++ scheme_generator/stdafx.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 
 //////////////////////////////////////////////////////////////////////////
-_loggerptr               g_pLogger = NULL;
-_loggerptr               g_pNetlogger = NULL;
+_loggerptr               g_pLogger = (int)NULL;
+_loggerptr               g_pNetlogger = (int)NULL;
 CDBWrapper               g_dbWrapper;
 CNetManager              g_netManager;
